HomeWork/Kỳ 3/DDM501.7/Buổi Học/Buổi 3/3/tutorial_03/requirements.txt
scikit-learn==1.3.2
mlflow==2.17.2
pandas==2.1.1
numpy==1.23.5
matplotlib==3.8.0
seaborn==0.13.2
boto3==1.37.14
python-dotenv==1.0.1